requirements.txt
pycryptodome
cryptography
pillow